Completed Chapter 1 and Chapter 2 commit

+++ CH02/02_12/02_12-typecast1.c
@@ -1,13 +1,12 @@
 #include <stdio.h>
 
-int main()
-{
-	int a,b;
+int main() {
+  int a, b;
 
-	a = 100;
-	b = 7;
+  a = 100;
+  b = 7;
 
-	printf("%d / %d = %d\n",a,b,a/b);
+  printf("%d / %d = %f\n", a, b, (float)a / b);
 
-	return(0);
+  return (0);
 }
